+++ frontend/app/src/components_mobile/home/wallet/SendToAddress.svelte
@@ -26,7 +26,7 @@
         type ResourceKey,
     } from "@client";
     import { isAccountIdentifierValid, isICRCAddressValid, publish } from "@shared";
-    import { getContext, onMount } from "svelte";
+    import { getContext, onDestroy, onMount } from "svelte";
     import { _ } from "svelte-i18n";
     import Account from "svelte-material-icons/AccountBoxOutline.svelte";
     import ChevronRight from "svelte-material-icons/ChevronRight.svelte";
@@ -148,6 +148,8 @@
             getCkbtcMinterWithdrawalInfo(0n);
         }
     });
+
+    onDestroy(() => ckbtcMinterInfoDebouncer.cancel());
 
     // Whenever the networks list changes, autoselect the first one
     $effect(() => {